Assignment complete
Grid system still had some issues with the images but I think that's
largely due to images that I selected and the way pure formats them. All
in all, requirements are met. It could always be better but I think it's
good enough for this assignment. I need to validate my work more often
and address tiny errors faster